Suppression de l'action findDossier
